requirements.txt
absl-py==0.7.1
astor==0.8.0
backcall==0.1.0
certifi==2019.6.16
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
keras-bert==0.70.1
keras-contrib==2.0.8
keras-embed-sim==0.7.0
keras-layer-normalization==0.12.0
keras-multi-head==0.20.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.0
keras-self-attention==0.41.0
keras-transformer==0.29.0
Markdown==3.1.1
numpy==1.16.4
pandas==0.25.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0rc1
tensorflow-gpu==1.14.0
tensorflow-hub==0.4.0
termcolor==1.1.0
tqdm==4.32.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.1
